# Improve scan performance and correctness with a prefix tree